HibernateTest: min ref

+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/SpendJdbcExtension.java
@@ -32,6 +32,7 @@ public class SpendJdbcExtension extends AbstractSpendExtension {
 
     @Override
     protected void removeSpend(SpendJson spend) {
+        if (spend == null) return;
         spendRepository.removeSpend(spend);
     }
 }
